Cambio de estructura de clases + Optimización
